standardizing return objects to work with all of the middleware, currency conversion, and allowable time period functions. Encountering issue with returned object on /product/pricebymarket route
