+++ time-filter/index.js
@@ -1,7 +1,7 @@
 
 
 module.exports = async (data, allowedPeriod ) =>{
-    console.log(`data time filter `, data)
+    // console.log(`data time filter `, data)
 
     //create date from most recent record in returned data
     const recentDate = new Date(data.recentRecordDate)
